+++ src/main/java/com/transgate/api/interfaces/WalletsInterface.java
@@ -5,6 +5,7 @@
  */
 package com.transgate.api.interfaces;
 
+import java.math.BigDecimal;
 import org.springframework.http.ResponseEntity;
 
 /**
@@ -21,7 +22,7 @@ public interface WalletsInterface {
     
     public ResponseEntity GetWalletActivity(String walletnumber);
 
-    public ResponseEntity InitiateDebitCreditWallet(String sessiontoken, String walletnumber, String actionType, double amount, String fundby);
+    public ResponseEntity InitiateDebitCreditWallet(String sessiontoken, String walletnumber, String actionType, BigDecimal amount, String fundby);
     
     public ResponseEntity MapWalletToUser(String sessiontoken, String walletnumber, String assignee, String username);
